split huge "tags" module into one module per tag
